Support predefined accounts during checkout.
Previously, only the standard account credentials (those stored in the database) were passed to `camptix-stripe.js` via `CampTixStripeData`. If a predefined account was used, then the values for the public and secret keys would be empty, and the transaction would fail.

The way that predefined accounts are handled in CampTix and the PayPal addon is somewhat messy, since there is the `$this->options` variable, which contains the options from the database, and a local `$options` value, which contains the predefined accounts. This uses the same fundamental approach of a separate variable, but centralizes the logic in a single function, rather than having it duplicated like the PayPal addon does.